+++ src/cards/components/sensor-collection/styles.ts
@@ -1,0 +1,44 @@
+import { css } from 'lit';
+
+/**
+ * Sensor styles for individual sensor display
+ * Includes layout, icon visibility, and sensor-specific styles
+ */
+export const styles = css`
+  :host {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 4px;
+    opacity: var(--text-opacity-theme, 0.4);
+    margin-left: -2%;
+    margin-top: 1%;
+  }
+
+  :host([hide-icons]) {
+    column-gap: 8px;
+    margin-left: 0px;
+  }
+
+  :host([layout='stacked']) {
+    flex-direction: column;
+  }
+
+  :host([layout='bottom']) {
+    position: absolute;
+    bottom: 2%;
+    left: 50%;
+    transform: translateX(-50%);
+  }
+
+  .sensor {
+    display: flex;
+    align-items: center;
+    gap: 4px;
+    --mdc-icon-size: var(--user-sensor-icon-size, 20px);
+  }
+
+  .sensor ha-state-icon,
+  .sensor ha-icon {
+    flex-shrink: 0;
+  }
+`;
