+++ test/cards/components/entity-collection.spec.ts
@@ -1,4 +1,4 @@
-import { EntityCollection } from '@cards/components/entity-collection';
+import { EntityCollection } from '@cards/components/entity-collection/entity-collection';
 import * as iconEntitiesModule from '@delegates/entities/icon-entities';
 import type { HomeAssistant } from '@hass/types';
 import * as iconModule from '@html/icon';
@@ -9,6 +9,7 @@ import type { EntityInformation, EntityState } from '@type/room';
 import { expect } from 'chai';
 import { html, nothing, type TemplateResult } from 'lit';
 import { stub } from 'sinon';
+import { styles } from '@cards/components/entity-collection/styles';
 
 describe('entity-collection.ts', () => {
   let element: EntityCollection;
@@ -267,9 +268,7 @@ describe('entity-collection.ts', () => {
 
   describe('styles', () => {
     it('should return entityStyles', () => {
-      // Import the actual styles to compare
-      const { entityStyles } = require('@theme/styles');
-      expect(EntityCollection.styles).to.equal(entityStyles);
+      expect(EntityCollection.styles).to.equal(styles);
     });
   });
 });
